project/src/requirements.txt
pyspark==3.1.1
python-dotenv==0.17.0
log4python==1.0.3